Address GeoJson/Polygon Layer audit
