+++ src/layers/core/primitive-polygon-layer/polygon-layer-64-vertex.glsl
