+++ src/layers/core/primitive-polygon-layer/polygon-layer-fragment.glsl
