+++ src/layers/core/primitive-polygon-layer/polygon-layer-vertex.glsl
