+++ src/layers/core/primitive-polygon-layer/polygon-tesselator-extruded.js
