+++ src/layers/core/primitive-polygon-layer/polygon-tesselator.js
