+++ src/layers/core/primitive-polygon-layer/polygon.js
